refacotr: migrated /api/menu endpoint
